\name{BaFTA-package}
\alias{BaFTA-package}
\alias{BaFTA}

\docType{package}

\title{
   Parametric Bayesian estimation of age-specific fertility for aggregated and individual level data
}

\description{
  This package makes inference on age-specific fertility trajectories from aggregated and individual data under a Bayesian framework. It includes tools for data checking, model diagnostics, and plotting.

}

\details{
\tabular{ll}{
	Package: \tab BaFTA \cr
	Type: \tab Package\cr
	Version: \tab 1.0.0\cr
	Date: \tab 2023-09-05\cr
	License: \tab GNU General Public Licence\cr
	LazyLoad: \tab yes\cr
}

Parametric estimation of age-specific fertility for aggregated and individual level data.
}

\author{Fernando Colchero \email{fernando_colchero@eva.mpg.de}}

\references{
Colchero, F. (\emph{In prep.}) Inference on age-specific fertility in ecology and evolution. Learning from other disciplines and improving the state of the art. 

Colchero, F., Eckardt, W. & Stoinski, T. (2021) Evidence of demographic buffering in an endangered great ape: Social buffering on immature survival and the role of refined sex-age classes on population growth rate. \emph{Journal of Animal Ecology} 90, 1701-1713.

Muller, M. N., Blurton Jones, N. G, Colchero, F., Thompson, M. E., Enigk, D. K. (2020) Sexual dimorphism in chimpanzee (Pan troglodytes schweinfurthii) and human age-specific fertility. \emph{Journal of human evolution} 144, 102795.

}